LIBFT

+++ ft_putnbr_fd.c
@@ -3,7 +3,7 @@
 
 void	ft_putnbr_fd(int n, int fd)
 {
-	long long nbr;
+	long int	nbr;
 
 	nbr = n;
 	if (nbr < 0)
